+++ server/controllers/Vehicles.js
@@ -1,6 +1,10 @@
+// This is the controller which works with the vehicles table in the database
 import Vehicle from "../models/vehicle.js";
 import Image from "../models/image.js";
+import Order from "../models/order.js";
+import sequelize from "../utils/database.js";
 
+// A function which fethces all the vehicles from the database
 const fetchVehicles = async (req, res, next) => {
   await Vehicle.findAll({
     include: [{ model: Image, as: "images" }],
@@ -13,6 +17,7 @@ const fetchVehicles = async (req, res, next) => {
   });
 };
 
+// A function which adds a new vehicle to the vehicles tablee
 const addVehicle = async (req, res, next) => {
   await Vehicle.create({
     title: req.body.title,
@@ -34,11 +39,9 @@ const addVehicle = async (req, res, next) => {
       console.log(err);
       res.status(502).json({ message: "error while adding the vehicle" });
     });
-  // const id = dbAttraction.id;
-  // updateImages(req.body.images, id);
-  // console.log(dbAttraction);
 };
 
+// A function which adds or updates the images of the vehicle
 const addImages = async (images, vehicleId) => {
   await Image.destroy({
     where: {
@@ -56,6 +59,7 @@ const addImages = async (images, vehicleId) => {
   });
 };
 
+// A function which updates the vehicle info
 const updateVehicle = async (req, res, next) => {
   const vehicleToUpdate = await Vehicle.findOne({
     where: { title: req.body.title },
@@ -88,29 +92,46 @@ const updateVehicle = async (req, res, next) => {
     });
 };
 
+// A function which deletes a vehicle from the database
 const deleteVehicle = async (req, res, next) => {
-  await Image.destroy({
+  // The vehicle cannot be deleted if it does exist in a "pending" order
+  await Order.findOne({
     where: {
       vehicleId: req.body.id,
+      status: "Pending",
     },
-  });
+  }).then(async (vehicleOrder) => {
+    if (vehicleOrder) {
+      return res.status(409).json({
+        message:
+          "This vehicle cannot be deleted at this moment, there is an open reservation for this vehicle.",
+      });
+    } else {
+      await Image.destroy({
+        where: {
+          vehicleId: req.body.id,
+        },
+      });
 
-  await Vehicle.destroy({
-    where: {
-      id: req.body.id,
-    },
-  })
-    .then(() => {
-      res.status(200).json({ message: "Vehicle deleted successfully" });
-    })
-    .catch((err) => {
-      console.log(err);
-      res
-        .status(502)
-        .json({ message: "An error occured while deleting the vehicle" });
-    });
+      await Vehicle.destroy({
+        where: {
+          id: req.body.id,
+        },
+      })
+        .then(() => {
+          res.status(200).json({ message: "Vehicle deleted successfully" });
+        })
+        .catch((err) => {
+          console.log(err);
+          res
+            .status(502)
+            .json({ message: "An error occured while deleting the vehicle" });
+        });
+    }
+  });
 };
 
+// A function which fetches a vehicle from the database by it's id
 const fetchVehicleById = async (req, res, next) => {
   await Vehicle.findByPk(req.body.id, {
     include: [{ model: Image, as: "images" }],
@@ -119,10 +140,60 @@ const fetchVehicleById = async (req, res, next) => {
   });
 };
 
+// A function which increments the quantity of a vehicle, after completing the same vehicle's orders
+const incrementQuantity = async (req, res, next) => {
+  const vehicleToUpdate = await Vehicle.findOne({
+    where: {
+      title: req.body.title,
+    },
+  });
+  if (vehicleToUpdate) {
+    await Vehicle.update(
+      {
+        quantity: vehicleToUpdate.dataValues.quantity + 1,
+      },
+      {
+        where: {
+          id: vehicleToUpdate.dataValues.id,
+        },
+      }
+    )
+      .then(() => {
+        res
+          .status(200)
+          .json({ message: "Vehicle quantity successfully updated" });
+      })
+      .catch((err) => {
+        console.log(err);
+        res
+          .status(502)
+          .json({ message: "An error updating the vehicle quantity" });
+      });
+  }
+};
+
+// A function which updates the status of a vehicle to "unavailable" when the available quantity is down to zero
+const unavailableVehicles = async (req, res, next) => {
+  await sequelize
+    .query(
+      "UPDATE vehicles SET available=CASE WHEN quantity=0 THEN 0 WHEN quantity>0 THEN 1 END;"
+    )
+    .then(() => {
+      res.status.json({
+        message: "Vehicles availability updated successfully ",
+      });
+    })
+    .catch((err) => {
+      console.log(err);
+    });
+};
+
 export {
   fetchVehicles,
   addVehicle,
   updateVehicle,
   deleteVehicle,
   fetchVehicleById,
+  incrementQuantity,
+  unavailableVehicles,
 };
